changes regarding linter rules

+++ lib/samples/sliders/slider/customization/shape_customization/slider_divisor_customization.dart
@@ -30,7 +30,7 @@ class _DivisorCustomizedSliderState extends SampleViewState {
       value: _value,
       onChanged: (dynamic values) {
         setState(() {
-          _value = values;
+          _value = values as double;
         });
       },
       interval: 10,
