+++ lib/samples/sliders/vertical_slider/customization/shape_customization/vertical_thumb_customization.dart
@@ -41,7 +41,7 @@ class _VerticalThumbCustomizedSliderState extends SampleViewState {
         value: _sliderValue,
         onChanged: (dynamic values) {
           setState(() {
-            _sliderValue = values;
+            _sliderValue = values as double;
           });
         },
         interval: 10,
